refactor(validator): rename to AuthKit namespace
