+++ AuthKit.slnx
@@ -12,7 +12,7 @@
     <Project Path="src/Plugins/Solutions/DevTokens/DevTokens.csproj" />
   </Folder>
   <Folder Name="/Tools/">
-    <Project Path="tools/PluginContractValidator/PluginContractValidator.csproj" />
+    <Project Path="tools/AuthKit.PluginContractValidator/AuthKit.PluginContractValidator.csproj" />
   </Folder>     
   <Project Path="src/Core/Core.csproj" />
   <Project Path="src/Host/Host.csproj" />
