+++ tools/AuthKit.PluginContractValidator/src/Rules/HealthRule.cs
@@ -1,9 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using AuthKit.PluginContractValidator.Core;
 using AuthKit.Plugins.Abstractions.Contracts;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using PluginContractValidator.Core;
 
-namespace PluginContractValidator.Rules;
+namespace AuthKit.PluginContractValidator.Rules;
 
 /// <summary>
 /// Ensures the plugin can configure its services and that
