--- tools/PluginContractValidator/PluginContractValidator.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net10.0</TargetFramework>
-    <RootNamespace>AuthKit.PluginContractValidator</RootNamespace>
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <ProjectReference Include="../../src/Plugins/Abstractions/AuthKit.Plugins.Abstractions.csproj" />
-  </ItemGroup>
-  
-  
-</Project>
